Synchronized databases and enabled internet checking

+++ app/src/main/java/com/example/foodplanner/View/FavoriteMealsAdapter.java
@@ -1,5 +1,6 @@
 package com.example.foodplanner.View;
 
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -8,17 +9,20 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.example.foodplanner.Controller.Activities.MainActivity;
 import com.example.foodplanner.Controller.Fragments.MainFragments.FavoriteMealsDirections;
 import com.example.foodplanner.Model.MealsItem;
 import com.example.foodplanner.Model.Repository;
 import com.example.foodplanner.Model.RootSingleMeal;
 import com.example.foodplanner.R;
+import com.example.foodplanner.Utility.NetworkChecker;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
@@ -37,6 +41,8 @@ public class FavoriteMealsAdapter extends RecyclerView.Adapter<FavoriteMealsAdap
     Observable<RootSingleMeal> observableMealSelectedFromFavorites;
 
     Repository rep;
+
+    private ProgressDialog progressDialog;
 
 
 
@@ -60,81 +66,75 @@ public class FavoriteMealsAdapter extends RecyclerView.Adapter<FavoriteMealsAdap
     @Override
     public void onBindViewHolder(@NonNull FavoriteMealsAdapter.ViewHolder holder, int position) {
 
-        /* Favorites Firestore part 3/4: Loading in recycler view and Removing */
-/*
         MealsItem mealsItem = mealsFavorite.get(position);
         holder.fav_tv_mealName.setText(mealsItem.getStrMeal());
         holder.fav_tv_mealArea.setText(mealsItem.getStrArea());
 
         Glide.with(viewGroup.getContext()).load(mealsItem.getStrMealThumb()).into(holder.fav_img_mealImg);
+
+        /* Favorites Firestore + Room part 3/4: Loading in recycler view and Removing */
+
+        NetworkChecker networkChecker = NetworkChecker.getInstance();
+
+
+
+
         holder.btn_removeFromFavorites.setOnClickListener(new View.OnClickListener() {        //\\\\\\\
             @Override
             public void onClick(View view) {
-                FirebaseFirestore.getInstance().collection("userFavorites").document(mealsItem.documentID)
-                        .delete()
-                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void aVoid) {
-                                Log.i(TAG, "DocumentSnapshot successfully deleted!");
-                                //(FavoriteMealsAdapter.this).notifyDataSetChanged();
-                                mealsFavorite.remove(position);
-                                notifyDataSetChanged();
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Log.i(TAG, "Error deleting document", e);
-                            }
-                        });                                            //\\\\\\\
-            }
-        });
-
-        holder.itemView.setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View view) {
 
 
-                Navigation.findNavController(viewGroup).navigate(FavoriteMealsDirections.actionNavFavoriteMealsToMealDeatailsFragment(mealsFavorite.get(position)));
-                mealsFavorite.clear(); //cause clicked back multiplied whats shown in the view.
+                if(!networkChecker.checkIfInternetIsConnected()){
+                    MainActivity.mainActivity.runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            Toast.makeText(MainActivity.mainActivity, "Turn internet on to be able to remove meals from your favorites..", Toast.LENGTH_SHORT).show();
+                        }
+                    });
 
-            }
-        });
+                } else if (networkChecker.checkIfInternetIsConnected()){
 
- */
+                    progressDialog = new ProgressDialog(viewGroup.getContext());
+                    progressDialog.setTitle("Removing meal from favorites");
+                    progressDialog.setMessage("Please wait while removing the selected item from your favorites.");
+                    progressDialog.setCanceledOnTouchOutside(false);
+                    progressDialog.show();
 
 
 
-        /* Favorites Room part 3/4:  Loading in recycler view and Removing */
+                    FirebaseFirestore.getInstance().collection("userFavorites").document(mealsItem.documentID)
+                            .delete()
+                            .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                @Override
+                                public void onSuccess(Void aVoid) {
+                                    Log.i(TAG, "DocumentSnapshot successfully deleted!");
+                                    //(FavoriteMealsAdapter.this).notifyDataSetChanged();
 
+                                    rep=new Repository(viewGroup.getContext());
+                                    rep.delete(mealsItem);
+                                    mealsFavorite.remove(position);
+                                    notifyDataSetChanged();
 
+                                    progressDialog.dismiss();
 
-
-            MealsItem mealsItem = mealsFavorite.get(position);
-            holder.fav_tv_mealName.setText(mealsItem.getStrMeal());
-            holder.fav_tv_mealArea.setText(mealsItem.getStrArea());
-
-            Glide.with(viewGroup.getContext()).load(mealsItem.getStrMealThumb()).into(holder.fav_img_mealImg);
-
-            holder.btn_removeFromFavorites.setOnClickListener(new View.OnClickListener() {        //\\\\\\\
-                @Override
-                public void onClick(View view) {
-
-                    rep=new Repository(viewGroup.getContext());
-                    rep.delete(mealsItem);
-                    mealsFavorite.remove(position);
-                    notifyDataSetChanged();
-
+                                }
+                            })
+                            .addOnFailureListener(new OnFailureListener() {
+                                @Override
+                                public void onFailure(@NonNull Exception e) {
+                                    Log.i(TAG, "Error deleting document", e);
+                                }
+                            });
                 }
-            });
 
+
+            }
+        });
 
         holder.itemView.setOnClickListener(new View.OnClickListener() {
 
             @Override
             public void onClick(View view) {
-
                 Navigation.findNavController(viewGroup).navigate(FavoriteMealsDirections.actionNavFavoriteMealsToMealDeatailsFragment(mealsFavorite.get(position)));
                 mealsFavorite.clear(); //cause clicked back multiplied whats shown in the view.
 
@@ -146,14 +146,8 @@ public class FavoriteMealsAdapter extends RecyclerView.Adapter<FavoriteMealsAdap
     @Override
     public int getItemCount() {
 
-        /* Favorites Firestore part 4/4: Getting item count */
-        /*
-         return mealsFavorite.size();
+        /* Favorites Firestore + Room part 4/4: Getting item count */
 
-         */
-
-
-         /* Favorites Room part 4/4: Getting item count */
         return mealsFavorite.size();
